+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added new `compas_timber._fabrication.StepJointNotch`.
 * Added new `compas_timber.connections.TStepJoint`.
 * Added new `utilities` module in `connections` package.
+* Added new `fasteners` module with new `Fastener` element type.
 
 ### Changed